zakladni uklid nad UzivatelPresenterem
