+++ app/presenters/BasePresenter.php
@@ -71,7 +71,7 @@ abstract class BasePresenter extends Nette\Application\UI\Presenter
             $ap = $ipIsInAp->fetch();
             $this->redirect('Ap:show', array('id'=>$ap->id));
         }
-        $this->redirect('Uzivatel:listall', array('search' => $values->search, 'id' => null));
+        $this->redirect('UzivatelList:listall', array('search' => $values->search, 'id' => null));
     }
 
     protected function getSubnetLinkFromIpAddress($ipAddress) {
