+++ app/templates/UzivatelList/listall.latte
@@ -7,9 +7,9 @@
 {/if}
 &nbsp;&nbsp;
 {if $canViewOrEdit}
-<a n:if="!$money" href="{link Uzivatel:listall search=>$search, money=>1}" class="btn btn-default btn-sm" role="button">Informace z money</a>
+<a n:if="!$money" href="{link UzivatelList:listall search=>$search, money=>1}" class="btn btn-default btn-sm" role="button">Informace z money</a>
 {/if}
-<a n:if="$money" href="{link Uzivatel:listall search=>$search}" class="btn btn-default btn-sm" role="button">Bez informací z money</a>
+<a n:if="$money" href="{link UzivatelList:listall search=>$search}" class="btn btn-default btn-sm" role="button">Bez informací z money</a>
 
 <br><br>
